+++ src/components/searchBar/searchBar.test.tsx
