+++ src/pages/MainPage/MainPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import SearchBar from './../components/searchBar';
-import CardsWrapper from './../components/cardsWrapper';
+import SearchBar from './../../components/searchBar/searchBar';
+import CardsWrapper from './../../components/cardsWrapper/cardsWrapper';
 
 function MainPage() {
   return (
